minimal config of matritools package
