+++ matritools/nodefilerow.py
@@ -1,453 +1,4 @@
-# Last updated 7/9/21 : 4:22PM PST
-
 import pandas as pd
-import numbers
-from ast import literal_eval
-import json
-import copy
-
-# ***********************
-# Commonly used functions
-# ***********************
-
-#region commonly used functions
-
-#region data exploration
-
-def print_df_column_set(df, column, print_sets=False):
-    """
-    Prints every out the number unique values in a dataframe column
-    :param df: target data frame
-    :param column: target column
-    :param print_sets: Set True to print out each of the unique values
-    :return: None
-    """
-    col_list = df[column].tolist()
-    col_set = set(col_list)
-
-    if print_sets:
-        print("\n\t" + column + ", \tLength: " + str(len(col_set)) + "\n")
-        for value in col_set:
-            print("\t\t" + str(value))
-    else:
-        print("\t" + column + ", \tLength: " + str(len(col_set)))
-
-def print_df_column_min_max(df, min_max_columns):
-    """
-    Prints out the min and make of each column in a data frame
-    :param df: target data frame
-    :param min_max_columns: list of target columns
-    :return: None
-    """
-    for column in min_max_columns:
-        if issubclass(df[column].dtype.type, numbers.Integral):
-            print("\t" + column + ", \tMin: " + str(df[column].min()) + ", \tMax: " + str(df[column].max()))
-
-def explore_df(df, set_columns=None, min_max_columns=None, print_sets=False):
-    """
-    Prints out the number of unique values from a list of columns for a data frame as well as the min and max values of each column
-    :param df: target dataframe
-    :param set_columns: list of columns to display number of unique values (default None)
-    :param min_max_columns: list of columns to display min and max values (default None)
-    :param print_sets: Set True to print out each of the unique values of the given columns (default False)
-    :return:
-    """
-
-    print("Shape: " + str(df.shape))
-    __explore_sets__(df, set_columns, print_sets)
-
-    if min_max_columns is not None:
-        print("\nIntergral Min/Max:")
-        print_df_column_min_max(df, min_max_columns)
-
-def __explore_sets__(df, set_columns, print_sets):
-    """
-    Helper function that calls print_df_column_set as well as a label depending on value of print_sets
-    :param df: target dataframe
-    :param set_columns: list of columns to display number of unique values
-    :param print_sets: Set True to print out each of the unique values of the given columns
-    :return:
-    """
-    if set_columns is None:
-        return
-    if print_sets:
-        print("Sets:")
-    else:
-        print("Set Lengths:")
-    for column in set_columns:
-        print_df_column_set(df, column, print_sets)
-
-#endregion
-
-def create_df_from_json(json_file_name):
-    with open(json_file_name) as json_data:
-        return pd.DataFrame(json.load(json_data))
-
-
-def create_df_from_json_string(json_string):
-    return pd.DataFrame(json.load(json_string))
-
-#region data manipulation
-
-def interpolate_df_column(df, column, obj_scale_min, obj_scale_max, column_tail):
-    """
-    Scales the values of a data frame column between obj_scale_min and obj_scale_max.
-    Adds the interpolated data into a new column labeled (original column name) + (column_tail) and preserves original
-    un-interpolated data.
-
-    :param df:
-    :param column:
-    :param obj_scale_min:
-    :param obj_scale_max:
-    :param column_tail:
-    :return:
-    """
-    col_min = df[column].min()
-    col_max = df[column].max()
-    col_list = df[column].tolist()
-    scalar = make_interpolater(col_min + .0002, col_max, obj_scale_min, obj_scale_max)
-    col_interp = [scalar(x) for x in col_list]
-    df[(column + column_tail)] = col_interp
-
-
-def make_interpolater(old_min, old_max, new_min, new_max):
-    """
-    Creates a reusable interpolation function to scale a value in between a new min and new max
-
-    :param old_min:
-    :param old_max:
-    :param new_min:
-    :param new_max:
-    :return: interpolation function
-    """
-
-    # Figure out how 'wide' each range is
-    old_span = old_max - old_min
-    new_span = new_max - new_min
-
-    # Compute the scale factor between old and new values
-    scale_factor = float(new_span) / float(old_span)
-
-    # create interpolation function using pre-calculated scaleFactor
-    def interp_fn(value):
-        """
-        Scale a value in between a set min and  max
-        :param value: value to be interpolated
-        :return: float
-        """
-        return new_min + (value - old_min) * scale_factor
-
-    return interp_fn
-
-
-
-def separate_compound_dataframe(df, evaluate_df=True, name_template=""):
-    """
-    Use to create a list of data frames from a dataframe with embedded lists.
-
-    :param df: a data frame that has lists of data contained in individual cells
-    :param evaluate_df: set True to apply literal_eval (default True)
-    :param name_template: individual data frame columns = lambda x: (name_template + "{}").format(x + 1) (default "")
-    :return: list of data frames
-
-
-    I.E when a data frame is in the form of :
-
-    [ [ 11, 12, 13], [ 14, 15, 16 ], [ 17, 18, 19 ] ]
-    [ [ 21, 22, 23], [ 24, 25, 26 ], [ 27, 28, 29 ] ]
-    [ [ 31, 32, 33], [ 34, 35, 36 ], [ 37, 38, 39 ] ]
-
-    returns
-
-    df1 [ 11, 12, 13]  df2 [ 14, 15, 16 ]  df3 [ 17, 18, 19 ]
-        [ 21, 22, 23]      [ 24, 25, 26 ]      [ 27, 28, 29 ]
-        [ 31, 32, 33]      [ 34, 35, 36 ]      [ 37, 38, 39 ]
-    """
-    # Expand evaluate tupled data into lists
-    if evaluate_df:
-        for column in df.columns:
-            df[column] = df[column].apply(literal_eval)
-
-    df_list = []
-
-    l = lambda x: (name_template + "{}").format(x + 1)
-    for column in df.columns:
-        df_list.append(pd.DataFrame(df[column].values.tolist(), df.index, dtype=object).fillna(17).rename(columns=l))
-
-    return df_list
-
-#endregion
-
-#region utility
-
-def find_index_in_set(value_set, value):
-    """
-    returns the index of a value in a set as if it were a list
-    :param value_set: target set
-    :param value: target value
-    :return: (int) index of value in value_set, returns -1 if value is not in value_set
-    """
-    index = 0
-    for i in value_set:
-        if value == i:
-            return index
-        index += 1
-    return -1
-
-def set_to_list(value_set):
-    """
-    Converts a set to a list
-    :param value_set: target set
-    :return: list
-    """
-    result = []
-    for value in value_set:
-        result.append(value)
-    return result
-
-#endregion
-#endregion
-
-
-
-
-# *****************
-# Node File Classes
-# *****************
-
-class NodeFile:
-    """
-    Virtual representation of an ANTZ node csv
-
-    Used to construct a virtual antz node file.
-    On construction, provide a file name. The constructor will append "_node.csv" to the end of the provided file name.
-
-    Basic usage:
-      After created node_file class, set the properties of the class to what you want a node csv file row to be,
-      then call create_node_file_row.
-
-      After adding all desired rows, call write to csv.
-
-    Advanced usage:
-      Create an AntzGlyph, modify its rows and use NodeFile.add_glyph_to_rows.
-      Then call AntzGlyph.increment_glyph, modify rows again accordingly, then use NodeFile.add_glyph_to_rows.
-      Repeat as necessary.
-    """
-
-    taginc = 1  # line number for a tag file row
-
-    header = "id,type,data,selected,parent_id," \
-             "branch_level,child_id,child_index,palette_id,ch_input_id," \
-             "ch_output_id,ch_last_updated,average,samples,aux_a_x," \
-             "aux_a_y,aux_a_z,aux_b_x,aux_b_y,aux_b_z," \
-             "color_shift,rotate_vec_x,rotate_vec_y,rotate_vec_z,rotate_vec_s," \
-             "scale_x,scale_y,scale_z,translate_x,translate_y," \
-             "translate_z,tag_offset_x,tag_offset_y,tag_offset_z,rotate_rate_x," \
-             "rotate_rate_y,rotate_rate_z,rotate_x,rotate_y,rotate_z," \
-             "scale_rate_x,scale_rate_y,scale_rate_z,translate_rate_x,translate_rate_y," \
-             "translate_rate_z,translate_vec_x,translate_vec_y,translate_vec_z,shader," \
-             "geometry,line_width,point_size,ratio,color_index," \
-             "color_r,color_g,color_b,color_a,color_fade," \
-             "texture_id,hide,freeze,topo,facet," \
-             "auto_zoom_x,auto_zoom_y,auto_zoom_z,trigger_hi_x,trigger_hi_y," \
-             "trigger_hi_z,trigger_lo_x,trigger_lo_y,trigger_lo_z,set_hi_x," \
-             "set_hi_y,set_hi_z,set_lo_x,set_lo_y,set_lo_z," \
-             "proximity_x,proximity_y,proximity_z,proximity_mode_x,proximity_mode_y," \
-             "proximity_mode_z,segments_x,segments_y,segments_z,tag_mode," \
-             "format_id,table_id,record_id,size\n"
-
-    node_file_rows = []
-
-    def __init__(self, file_name):
-        """
-        :param file_name: name of node and tag file created on write_to_csv()
-        """
-        self.properties = NodeRowProperties()
-        self.node_file_name = file_name
-        self.__add_initial_rows__()
-
-    def get_row_by_index(self, index):
-        return self.node_file_rows[index]
-
-    def get_row_by_id(self, row_id):
-        for row in self.node_file_rows:
-            if row.properties.id == row_id:
-                return row
-
-    def write_to_csv(self):
-        """ Write all node_file_rows to csv. This will overwrite the file."""
-        node_file = open(self.node_file_name + "_node.csv", "w")
-        tag_file = open(self.node_file_name + "_tag.csv", "w")
-        tag_file.write("id,record_id,table_id,title,description\n")
-
-        node_file.write(self.header)
-
-        for file_row in self.node_file_rows:
-            node_file.write(file_row.properties.to_string())
-            if file_row.properties.tag_text != "":
-                tag_text = str(self.taginc) + "," + \
-                           str(file_row.properties.record_id) + ",0,\"" + \
-                           str(file_row.properties.tag_text) + "\",\"\"\n"
-
-                self.taginc += 1
-                tag_file.write(tag_text)
-
-        node_file.close()
-
-    def create_node_row(self, tag_text=""):
-        """
-        Creates a node file row from node_file current properties and increments id by 1
-        :param tag_text: text that will be written in the tag file associated with this node file row (default "")
-        :return: None
-        """
-        node_row = NodeFileRow(self.properties.to_string())
-        self.node_file_rows.append(node_row)
-        if tag_text != "":
-            node_row.properties.tag_text = tag_text
-        self.properties.set_id(self.properties.id + 1)
-
-    def add_glyph_to_node_file_rows(self, glyph):
-        """
-        Appends all rows of a glyph to node_file_rows and increments glyph template id, parent_id, data, and record_id
-        :param glyph:
-        :return: None
-        """
-        for row in glyph.node_file_rows:
-            self.node_file_rows.append(copy.deepcopy(row))
-        glyph.increment_node_file_rows()
-
-    def to_dataframe(self):
-        """ returns a data frame with all node file rows """
-        list_of_lists = []
-        for row in self.node_file_rows:
-            columns = []
-            for column in row.properties.to_string().split(","):
-                columns.append(column)
-            list_of_lists.append(columns)
-        return pd.DataFrame(list_of_lists)
-
-    def __add_initial_rows__(self):
-        # Row for world parameters
-        self.node_file_rows.append(NodeFileRow("1,0,1,0,0,"
-                                               "0,1,0,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,1,"
-                                               "1,1,1,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,1,0,0.1,0,"
-                                               "50,101,101,255,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,16,16,0,0,"
-                                               "0,0,0,420"))
-        # Row for first camera view
-        self.node_file_rows.append(NodeFileRow("2,1,2,0,0,"
-                                               "0,2,2,3,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0.008645,0.825266,-0.564678,"
-                                               "1,1,1,-32.446629,-180.908295,"
-                                               "143.514175,0,0,1,0,"
-                                               "0,0,55.620094,0.6002,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "1,0,0.1,0,"
-                                               "50,101,101,255,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "214.306686,0,0,0,0,"
-                                               "0,16,16,0,0,"
-                                               "0,0,0,420"))
-        # Row for second camera view
-        self.node_file_rows.append(NodeFileRow("3,1,3,0,2,"
-                                               "1,3,0,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,-1,"
-                                               "1,1,1,-0.5,0,"
-                                               "571.75,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,1,0,0.1,0,"
-                                               "50,101,101,255,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,16,16,0,0,"
-                                               "0,0,0,420"))
-        # Third camera view
-        self.node_file_rows.append(NodeFileRow("4,1,4,0,2,"
-                                               "1,4,0,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,1,-0,"
-                                               "1,1,1,0,-90,"
-                                               "7,0,0,1,0,"
-                                               "0,0,90,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,1,0,0.1,0,"
-                                               "50,101,101,255,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,16,16,0,0,"
-                                               "0,0,0,420"))
-        # Fourth camera view
-        self.node_file_rows.append(NodeFileRow("5,1,5,0,2,"
-                                               "1,5,0,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,-1,-0,-0,"
-                                               "1,1,1,85,0,"
-                                               "7,0,0,1,0,"
-                                               "0,0,90,270,0,"
-                                               "0,0,0,0,0,"
-                                               "0,-0,0,0,0,"
-                                               "0,1,0,0.1,0,"
-                                               "50,101,101,255,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,16,16,0,0,"
-                                               "0,0,0,420"))
-        # Default Grid
-        self.node_file_rows.append(NodeFileRow("6,6,6,1,0,"
-                                               "0,0,1,0,0,"
-                                               "0,0,0,1,360,"
-                                               "220,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "1,1,1,0,0,"
-                                               "0,0,0,1,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,1,0,0.1,3,"
-                                               "0,0,255,150,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,0,0,0,0,"
-                                               "0,1,1,0,0,"
-                                               "0,0,0,420"))
-
 
 class NodeFileRow:
 
@@ -460,7 +11,7 @@ class NodeFileRow:
         if len(values) != 94:
             print("Values length = " + str(len(values)))
             print(comma_string)
-            raise RuntimeError("Comma seperated string has incorrect number of values")
+            raise RuntimeError("Comma separated string has incorrect number of values")
 
         self.set_properties_from_string_list(values)
 
@@ -560,94 +111,6 @@ class NodeFileRow:
         self.properties.record_id = int(float(values[92]))
         self.properties.size = int(float(values[93]))
 
-
-class AntzGlyph:
-    """
-    Used to represent a antz_glyph. Construct using node file generated from antz that contains only one glyph.
-    node_file_rows[0] is the root object.
-    """
-    node_file_rows = []
-
-    def __init__(self, csv_file_name):
-        """
-        :param csv_file_name: name of the glyph template csv file
-        """
-        if csv_file_name != "":
-            df = pd.read_csv(csv_file_name)
-            df = df.applymap(lambda cell: int(cell) if str(cell).endswith('.0') else cell)
-            for index, row in df.iterrows():
-                line = ""
-                for column in df.columns:
-                    line += str(row[column]) + ","
-                line = line[:len(line) - 1]
-                self.node_file_rows.append(NodeFileRow(line))
-        else:
-            raise RuntimeError("antz_glyph was constructed without a csv file name")
-
-    def increment_node_file_rows(self):
-        """
-        Use this to update the ids, parent ids, data, and record_id of each row of the glyph to represent a new glyph.
-        By default this gets called when adding glyph to node file.
-        :return: None
-        """
-        old_parent_ids = []
-        new_parent_ids = []
-
-        old_parent_ids.append(self.node_file_rows[0].properties.id)
-        row_id = self.node_file_rows[len(self.node_file_rows) - 1].properties.id
-        row_id += 1
-        new_parent_ids.append(row_id)
-        self.node_file_rows[0].properties.set_id(row_id)
-        self.node_file_rows[0].properties.tag_text = ""
-
-        for row in self.node_file_rows[1:]:
-            row_id += 1
-            row.properties.set_id(row_id)
-            row.properties.tag_text = ""
-
-            # find parent objects and give them updated parent ids
-            if row.properties.parent_id not in old_parent_ids:
-                old_parent_ids.append(row.properties.parent_id)
-                new_parent_ids.append(row_id - 1)
-                row.properties.parent_id = row_id - 1
-            else:
-                parent_id_index = self.__find_old_parent_id_index__(row, old_parent_ids)
-
-                row.properties.parent_id = new_parent_ids[parent_id_index]
-
-    def unselect_all(self):
-        """ Changes the selected property of all node file rows to 0. """
-        for row in self.node_file_rows:
-            row.properties.selected = 0
-
-    def match_record_ids_and_data_to_ids(self):
-        for row in self.node_file_rows:
-            row.properties.set_id(row.properties.id)
-
-    def get_rows_of_branch_level(self, branch_level):
-        """ Returns a list of NodeFileRow's of a given branch level"""
-        rows = []
-        for row in self.node_file_rows:
-            if row.properties.branch_level == branch_level:
-                rows.append(row)
-        return rows
-
-    def remove_rows_of_branch_level(self, branch_level):
-        """ removes all NodeFileRow's of a given branch level"""
-        rows = self.get_rows_of_branch_level(branch_level)
-
-        for row in rows:
-            self.node_file_rows.remove(row)
-
-
-    @staticmethod
-    def __find_old_parent_id_index__(row, old_parent_ids):
-        parent_id_index = 0
-        for row_id in old_parent_ids:
-            if row.properties.parent_id != old_parent_ids[parent_id_index]:
-                parent_id_index += 1
-            else:
-                return parent_id_index
 
 class NodeRowProperties:
     """ Container of properties and property setters for node file rows."""
@@ -1252,5 +715,3 @@ class NodeRowProperties:
         self.segments_x = x
         self.segments_y = y
         self.segments_z = z
-
-    # endregion
